+++ src/react/components/illustration/stories/variations.js
@@ -1,0 +1,34 @@
+const { sizeClasses } = require('../private/constants');
+
+const storiesWithVariationsKindName = 'Instances/XUIIllustration';
+
+const variations = [
+  {
+    storyKind: storiesWithVariationsKindName,
+    storyTitle: 'custom height',
+    height: 200,
+  },
+  {
+    storyKind: storiesWithVariationsKindName,
+    storyTitle: 'custom padding',
+    padding: 0,
+    size: 'medium',
+  },
+  {
+    storyKind: storiesWithVariationsKindName,
+    storyTitle: 'custom height and padding',
+    height: 200,
+    padding: 20,
+  },
+];
+
+const sizeVariations = Object.keys(sizeClasses).map(size => ({
+  storyKind: storiesWithVariationsKindName,
+  storyTitle: `${size} illustration`,
+  size,
+}));
+
+module.exports = {
+  storiesWithVariationsKindName,
+  variations: [...sizeVariations, ...variations],
+};