+++ src/react/components/pill/stories/variations.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import XUIAvatar from '../../avatar/XUIAvatar';
-
-const NOOP = require('../../helpers/noop');
+import NOOP from '../../helpers/noop';
 
 const storiesWithVariationsKindName = 'Instances/XUIPill';